requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.0.6
django-cors-headers==4.6.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
fastapi==0.115.6
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.2.2
mysqlclient==2.2.4
opcua==0.98.13
openpyxl==3.1.5
packaging==24.2
parsel==1.9.1
pillow==11.1.0
Protego==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyDispatcher==2.0.7
pymodbus==3.6.8
pyOpenSSL==24.3.0
pypdf==5.1.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.0
starlette==0.41.3
tldextract==5.1.3
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.34.0
w3lib==2.2.1
zope.interface==7.2
